refactor: 예약 삭제, 예약 대기 승인 메서드 분리
